requirements.txt
streamlit
tensorflow==2.15.0
numpy
pillow
streamlit-drawable-canvas
keras==2.15.0